requirements.txt
flask
python-dotenv
botbuilder-core
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-speech
openai
pydub